+++ OhmLab_FUHCM_BE/BusinessLayer/RequestModel/Assignment/GradeTeamMemberRequestModel.cs
@@ -4,7 +4,7 @@ namespace BusinessLayer.RequestModel.Assignment
 {
     public class GradeTeamMemberRequestModel
     {
-        public int IndividualGrade { get; set; }   // Điểm cá nhân (0-10)
+        public float IndividualGrade { get; set; }   // Điểm cá nhân (0-10)
         public string? IndividualComment { get; set; } // Nhận xét cá nhân
     }
 }
