+++ OhmLab_FUHCM_BE/BusinessLayer/ResponseModel/Assignment/TeamGradeResponseModel.cs
@@ -9,7 +9,7 @@ namespace BusinessLayer.ResponseModel.Assignment
         public string TeamName { get; set; } = null!;
         public int LabId { get; set; }
         public string LabName { get; set; } = null!;
-        public int TeamGrade { get; set; }         // Điểm chung
+        public double TeamGrade { get; set; }         // Điểm chung
         public string? TeamComment { get; set; }   // Nhận xét chung
         public List<TeamMemberGradeModel> Members { get; set; } = new List<TeamMemberGradeModel>();
         public DateTime? GradedDate { get; set; }
@@ -20,7 +20,7 @@ namespace BusinessLayer.ResponseModel.Assignment
     {
         public Guid StudentId { get; set; }
         public string StudentName { get; set; } = null!;
-        public int IndividualGrade { get; set; }   // Điểm cá nhân
+        public double IndividualGrade { get; set; }   // Điểm cá nhân
         public string? IndividualComment { get; set; } // Nhận xét cá nhân
     }
 }
